Arch: Put into arch.h
