+++ CMakeLists.txt
@@ -403,9 +403,9 @@ else()
     set_source_files_properties(${BISON_OUTPUT} PROPERTIES COMPILE_FLAGS "/wd4005 /wd4065")
 endif()
 
-# Set C++ standard to C++14.
+# Set C++ standard to C++17
 set_target_properties(${PROJECT_NAME} PROPERTIES
-    CXX_STANDARD 14
+    CXX_STANDARD 17
     CXX_STANDARD_REQUIRED YES)
 
 if (UNIX)
